requirements.txt
# Streamlit app
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Authentication
httpx-oauth>=0.13.0
firebase-admin>=6.0.0

# Core dependencies
pandas>=2.0.0
requests>=2.31.0
pyyaml>=6.0

# AI/ML
google-genai>=1.61.0

# Text-to-Speech (Google Cloud TTS)
google-cloud-texttospeech>=2.16.0

# Translation
translate>=3.6.1
beautifulsoup4>=4.12.0

# Image generation
Pillow>=10.0.0

# Database & Backend
firebase-admin>=6.0.0

# File handling
openpyxl>=3.0.0

# Anki deck generation
genanki>=0.13.0

# IPA/Phonetic transcription
epitran>=1.24

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.0

# Testing
pytest>=7.0.0
